+++ resources/views/device-tab/page.blade.php
@@ -1,0 +1,3 @@
+{{-- Section bar plus the selected section. Variables: DevicePageData::section(). --}}
+<x-option-bar name="NETCONF" :options="$sections" :selected="$section"></x-option-bar>
+@include('netconf::device-tab.' . $section)